+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### jannovar-core
 
+* changing treatment of insertions at exon/intron junctions; they are
+  considered as intronic insertions now that affect splicing
 * converting `GenomeVariant` of `AnnotationList` to always be on the forward
   strand after construction of  `AnnotationList`
 * deprecating the `{,All,Best}AnnotationTextGenerator` classes