+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## develop
 
+### jannovar-core
+
+* changing treatment of insertions at exon/intron junctions; they are
+  considered as intronic insertions now that affect splicing
+* converting `GenomeVariant` of `AnnotationList` to always be on the forward
+  strand after construction of  `AnnotationList`
+* deprecating the `{,All,Best}AnnotationTextGenerator` classes
+
+## v.0.13
+
 ### jannovar-cli
 
 * Moving `JannovarOptions` into jannovar-cli.